Make Term#recalculate_value define value in full
Using a calculator and assuming summation is less preferable because in some
cases summation may not be the desired outcome, and there is no need to
outsource the calculation when a subclasses are available.  This will allow
specific calculations to be optimized according to the algorithm; in other
words making Term the caching object that it wants to be.