Redefine Term methods to attach/detach single terms
Single methods will support delta-data a little better (where a one parent or
child update, vs a group of parent/children updates, will be the norm), as well
as making named parents/children easier to manage.  This algorithm may result
in walking the graph more times than previous -- if that turns out to be a time
sink then this approach could be revisited.