Spike out processing engine
Not factored right, but the basic idea is to have an engine with configurable
handlers.  Engines are responsible for the entry format; handlers are
responsible for converting entry data to the data used by the manager and
terms.  Currently a concise line format is used.

More features and refactoring to come.