Move definition of dimension types to controller
Placing them in the config file makes less sense; I anticipate it will be more
commmon to make a single controller designed to work with a fixed set of
dimension types vs making a generic controller designed for variable dimension
types.  This puts full control in the hands of the controller.